refactor(kmeans): split kmeans compile
