+++ src/algorithm/ivf_partition/gno_imi_partition.cpp
@@ -22,7 +22,7 @@
 #include <vector>
 
 #include "impl/allocator/safe_allocator.h"
-#include "impl/kmeans_cluster.h"
+#include "impl/cluster/kmeans_cluster.h"
 #include "inner_string_params.h"
 #include "utils/util_functions.h"
