+++ src/impl/cluster/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+add_library (cluster OBJECT kmeans_cluster.cpp)
+target_link_libraries (cluster PRIVATE fmt::fmt)
+maybe_add_dependencies (cluster spdlog openblas mkl)
+
